+++ src/components/feature/CardValidityPeriodInputField/constants.ts
@@ -1,0 +1,2 @@
+export const MONTH_MAX_LENGTH = 2;
+export const YEAR_MAX_LENGTH = 2;
