+++ src/components/feature/CardValidityPeriodInputField/utils.ts
@@ -1,0 +1,38 @@
+import {
+  checkIsInt,
+  validateMonthRange,
+  validateYearRange,
+} from "@/utils/validator";
+import type { ValidityPeriod } from "./CardValidityPeriodInputField";
+import { MONTH_MAX_LENGTH } from "./constants";
+import type { InputStatus } from "./errorMessage";
+
+export const checkCardNumberInputStatus = (
+  key: keyof ValidityPeriod,
+  value: string,
+): InputStatus => {
+  if (!checkIsInt(+value)) {
+    return "ONLY_NUMBER";
+  }
+
+  if (key === "month") {
+    if (value.length === 0) {
+      return "DEFAULT";
+    }
+
+    if (value.length === MONTH_MAX_LENGTH && !validateMonthRange(+value)) {
+      return "MONTH_RANGE_ERROR";
+    }
+    return "DEFAULT";
+  }
+  if (key === "year") {
+    if (value.length === 0) return "DEFAULT";
+    if (!validateYearRange(+value)) return "YEAR_RANGE_ERROR";
+    return "DEFAULT";
+  }
+  return "DEFAULT";
+};
+
+export const formatValidityPeriod = (nextRaw: string) => {
+  return nextRaw.replace(/\D/g, "").slice(0, 2);
+};
